requirements.txt
fastapi
sqlalchemy
alembic
psycopg2
uvicorn
pydantic
python-dotenv
python-jose
passlib
python-multipart